+++ src/links.rs
@@ -37,4 +37,5 @@ pub mod statuses {
     pub const SHOW: &'static str = "https://api.twitter.com/1.1/statuses/show.json";
     pub const RETWEETS_OF_STEM: &'static str = "https://api.twitter.com/1.1/statuses/retweets";
     pub const LOOKUP: &'static str = "https://api.twitter.com/1.1/statuses/lookup.json";
+    pub const HOME_TIMELINE: &'static str = "https://api.twitter.com/1.1/statuses/home_timeline.json";
 }